+++ tests/unit_tests/Commands/NewRelic/DisableCommandTest.php
@@ -44,17 +44,9 @@ class DisableCommandTest extends NewRelicCommandTest
 
         $this->new_relic->expects($this->once())
             ->method('disable')
-            ->with();
-        $this->site->expects($this->once())
-            ->method('converge')
             ->willReturn($workflow);
 
         $this->logger->expects($this->at(0))
-            ->method('log')->with(
-                $this->equalTo('notice'),
-                $this->equalTo('New Relic disabled. Converging bindings.')
-            );
-        $this->logger->expects($this->at(1))
             ->method('log')->with(
                 $this->equalTo('notice'),
                 $this->equalTo('successful workflow')